Refactor waypoint service into components
