+++ Assets/Scripts/Map/WaypointRegistry.cs
@@ -1,0 +1,23 @@
+using System.Collections.Generic;
+using System.Linq;
+
+public class WaypointRegistry
+{
+    private readonly Dictionary<RoomManager, List<RoomWaypoint>> roomWaypoints = new();
+
+    public void RegisterRoomWaypoints(RoomManager room, IEnumerable<RoomWaypoint> waypoints)
+    {
+        roomWaypoints[room] = waypoints.ToList();
+    }
+
+    public void UnregisterRoomWaypoints(RoomManager room)
+    {
+        roomWaypoints.Remove(room);
+    }
+
+    public List<RoomWaypoint> GetAllWaypoints() =>
+        roomWaypoints.Values.SelectMany(l => l).ToList();
+
+    public List<RoomWaypoint> GetActiveWaypoints() =>
+        roomWaypoints.Values.SelectMany(l => l).Where(wp => wp.IsAvailable).ToList();
+}
